chore: move manual scripts to scripts/manual and clean runtime artifacts
