+++ .github/workflows/run_tests.yml
@@ -34,4 +34,4 @@ jobs:
 
 
       - name: Run tests
-        run: python -m pytest --ignore=tests/scripts
+        run: python -m pytest --ignore=scripts/manual
