+++ .gitignore
@@ -185,3 +185,5 @@ examples/*
 render_logs/
 SimulationData/
 results/
+runtime.csv
+*-runtime.csv
