+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 testpaths = tests
-norecursedirs = tests/scripts
+norecursedirs = scripts/manual
 python_files = test_*.py
