--- runtime.csv
@@ -1,21 +1,0 @@
-Timesteps,Environment Creation,Agent Creation,Environment Reset,Render,Buildings Observations Retrieval,Total
-438,4.50667142868042,0.006586313247680664,4.519566535949707,5.535308122634888,1.6790244579315186,17.433846712112427
-876,4.56793212890625,0.006595611572265625,4.580786943435669,21.88694667816162,3.580393075942993,41.93264889717102
-1314,4.914121389389038,0.006538867950439453,4.926801681518555,45.75999712944031,5.24973201751709,73.2166600227356
-1752,4.504994869232178,0.006557941436767578,4.517852544784546,80.0065655708313,6.908231258392334,114.19591903686523
-2190,4.513540506362915,0.00652623176574707,4.526360988616943,123.26138281822205,8.714205265045166,165.21371245384216
-2628,4.544100046157837,0.011018991470336914,4.561450958251953,175.6818335056305,10.399182081222534,224.98745441436768
-3066,4.518075227737427,0.0066411495208740234,4.531431198120117,243.58430576324463,12.474161624908447,301.88921904563904
-3504,4.626904010772705,0.006839275360107422,4.640694856643677,311.51151871681213,14.237102031707764,378.1729738712311
-3942,4.5231921672821045,0.006663084030151367,4.536792039871216,397.3409082889557,16.430727005004883,473.9951739311218
-4380,4.829542636871338,0.006905317306518555,4.843914747238159,485.48155522346497,18.250758409500122,570.1552166938782
-4818,4.841669797897339,0.006783008575439453,4.856241941452026,594.1698787212372,20.284460067749023,687.6506559848785
-5256,4.7991251945495605,0.0068035125732421875,4.813369989395142,708.9323651790619,22.775211095809937,812.7648782730103
-5694,4.538013219833374,0.006818532943725586,4.552016496658325,807.7149128913879,23.798004627227783,916.2571015357971
-6132,4.890913009643555,0.007429838180541992,4.906832933425903,1015.7709510326385,27.666150093078613,1141.2058708667755
-6570,5.3231260776519775,0.00836801528930664,5.341875076293945,1125.5635533332825,29.228581190109253,1258.3709905147552
-7008,5.04076886177063,0.007013797760009766,5.056625604629517,1367.5648212432861,32.44820427894592,1514.6226062774658
-7446,5.265127420425415,0.007414817810058594,5.281713962554932,1527.6907942295074,34.36405372619629,1683.2380928993225
-7884,4.682467937469482,0.007017612457275391,4.6983325481414795,1525.358972787857,33.1479754447937,1675.5415420532227
-8322,4.7242751121521,0.007172346115112305,4.740380048751831,1690.7921800613403,34.96868872642517,1850.1829905509949
-8760,4.7104411125183105,0.006928443908691406,4.726260423660278,1875.0659368038177,37.01160001754761,2042.7379703521729
