+++ scripts/manual/bench_runtime_sweep.py
@@ -1,4 +1,4 @@
-import run_ev_rbc_export_end as obj
+import demo_ev_rbc_export_end as obj
 from pandas import DataFrame
 
 timesteps = [438, 876, 1314, 1752, 2190,
