+++ scripts/manual/diagnostics_occupant_override.py
