+++ scripts/manual/manual_compatibility_matrix.py
