+++ scripts/manual/manual_environment_smoke.py
