+++ scripts/manual/manual_lstm_smoke.py
