+++ scripts/manual/tacc_job.sh
